refactor: dissolve manager module
