+++ agent/subagents/worker/tools/fill_from_vault.ts
@@ -4,14 +4,18 @@ import { requireOwnedBrowserSession } from "@/agent/subagents/worker/lib/owned-b
 import { requireWorkerScope } from "@/agent/subagents/worker/lib/access";
 import { readVaultItem } from "@/db/services/vault";
 import { kernel } from "@/lib/kernel";
-import { materializeAutofillClaims } from "@/modules/manager/server/vault-autofill";
-import { vaultAutofillProvider } from "@/modules/manager/server/vault-autofill-provider";
 import {
   currentKernelPageOrigin,
   fillWithKernelNativeAutofill,
   nativeAutofillTokens,
-} from "@/modules/manager/server/kernel-native-autofill";
-import { fillFromVaultRequestSchema } from "@/modules/manager/vault-autofill";
+} from "../lib/autofill/native";
+import { vaultAutofillProvider } from "../lib/autofill/provider";
+import { materializeAutofillClaims } from "../lib/autofill/service";
+
+const inputSchema = z.object({
+  browserSessionId: z.string().trim().min(1).max(500),
+  candidateId: z.string().trim().min(1).max(500),
+});
 
 const outputSchema = z.object({
   filledClaims: z.number().int().nonnegative(),
@@ -23,7 +27,7 @@ const outputSchema = z.object({
 export default defineTool({
   description:
     "Fill a login, card, or address form with an opaque handle returned by list_vault. Focus one control in the intended form first. Never supply vault fields, selectors, origins, or secret values.",
-  inputSchema: fillFromVaultRequestSchema,
+  inputSchema,
   outputSchema,
   async execute(input, context) {
     const scope = await requireWorkerScope(context);
