+++ db/services/vault.ts
@@ -1,8 +1,28 @@
+import {
+  createCipheriv,
+  createDecipheriv,
+  randomBytes,
+  randomUUID,
+} from "node:crypto";
 import { and, desc, eq } from "drizzle-orm";
 import { z } from "zod";
+import {
+  loginAccountHint,
+  parsePaymentCardSecret,
+  parseLoginVaultPayload,
+  paymentCardBrand,
+  vaultItemKindSchema,
+  type VaultCreateItem,
+} from "@/lib/vault";
 import type { AccessScope } from "@/lib/access-scope";
-import { vaultItemKindSchema } from "@/modules/manager";
 import { db, vaultItems } from "@/db";
+import {
+  deleteEncryptedSecret,
+  readEncryptedSecret,
+  writeEncryptedSecret,
+} from "@/db/services/secrets";
+import { ensureScope } from "@/db/services/scope";
+import { env } from "@/lib/env";
 
 const vaultRecordSchema = z.object({
   account: z.string(),
@@ -24,7 +44,7 @@ const selection = {
   updatedAt: vaultItems.updatedAt,
 };
 
-export async function createVaultItem(scope: AccessScope, record: VaultRecord) {
+async function createVaultRecord(scope: AccessScope, record: VaultRecord) {
   await db.insert(vaultItems).values({
     ...record,
     workspaceId: scope.workspaceId,
@@ -41,6 +61,17 @@ export async function listVaultItems(scope: AccessScope) {
         .where(eq(vaultItems.workspaceId, scope.workspaceId))
         .orderBy(desc(vaultItems.updatedAt))
     );
+}
+
+export async function readVaultItems(scope: AccessScope) {
+  await ensureScope(scope);
+  const records = await listVaultItems(scope);
+  return Promise.all(
+    records.map(async (record) => ({
+      ...record,
+      hasSecret: await hasVaultSecret(scope, record.id),
+    }))
+  );
 }
 
 export async function readVaultItem(scope: AccessScope, id: string) {
@@ -61,5 +92,108 @@ export async function deleteVaultItem(scope: AccessScope, id: string) {
       and(eq(vaultItems.workspaceId, scope.workspaceId), eq(vaultItems.id, id))
     )
     .returning({ id: vaultItems.id });
-  return rows.length > 0;
+  if (rows.length === 0) return false;
+  await deleteEncryptedSecret(scope, id);
+  return true;
+}
+
+export async function saveVaultItem(
+  scope: AccessScope,
+  input: VaultCreateItem
+) {
+  await ensureScope(scope);
+  const id = randomUUID();
+  const now = new Date().toISOString();
+  await writeVaultSecret(scope, id, input.secret);
+
+  try {
+    await createVaultRecord(scope, {
+      account: vaultAccountHint(input),
+      createdAt: now,
+      id,
+      kind: input.kind,
+      label: input.label,
+      updatedAt: now,
+    });
+  } catch (error) {
+    await deleteEncryptedSecret(scope, id);
+    throw error;
+  }
+}
+
+export async function readVaultSecret(scope: AccessScope, id: string) {
+  const encrypted = await readEncryptedSecret(scope, id);
+  return encrypted ? decryptVaultSecret(scope, id, encrypted) : undefined;
+}
+
+export async function hasVaultSecret(scope: AccessScope, id: string) {
+  return (await readEncryptedSecret(scope, id)) !== undefined;
+}
+
+async function writeVaultSecret(scope: AccessScope, id: string, value: string) {
+  await writeEncryptedSecret(scope, id, encryptVaultSecret(scope, id, value));
+}
+
+function vaultAccountHint(input: VaultCreateItem) {
+  switch (input.kind) {
+    case "login": {
+      const payload = parseLoginVaultPayload(input.secret);
+      if (!payload)
+        throw new Error("The saved login is incomplete or invalid.");
+      return loginAccountHint(
+        payload.identifier,
+        "origin" in payload ? payload.origin : undefined
+      );
+    }
+    case "payment": {
+      const card = parsePaymentCardSecret(input.secret);
+      return `${paymentCardBrand(card.number)} · •••• ${card.number.slice(-4)}`;
+    }
+    case "address":
+    case "contact":
+      return "";
+  }
+}
+
+function encryptVaultSecret(scope: AccessScope, id: string, value: string) {
+  const iv = randomBytes(12);
+  const cipher = createCipheriv(
+    "aes-256-gcm",
+    Buffer.from(env.SECRET_ENCRYPTION_KEY, "base64"),
+    iv
+  );
+  cipher.setAAD(vaultSecretAad(scope, id));
+  const ciphertext = Buffer.concat([
+    cipher.update(value, "utf8"),
+    cipher.final(),
+  ]);
+  return [
+    "v1",
+    iv.toString("base64url"),
+    cipher.getAuthTag().toString("base64url"),
+    ciphertext.toString("base64url"),
+  ].join(".");
+}
+
+function decryptVaultSecret(scope: AccessScope, id: string, value: string) {
+  const [version, encodedIv, encodedTag, encodedCiphertext] = value.split(".");
+  if (version !== "v1" || !encodedIv || !encodedTag || !encodedCiphertext) {
+    throw new Error("The stored secret uses an unsupported format.");
+  }
+
+  const decipher = createDecipheriv(
+    "aes-256-gcm",
+    Buffer.from(env.SECRET_ENCRYPTION_KEY, "base64"),
+    Buffer.from(encodedIv, "base64url")
+  );
+  decipher.setAAD(vaultSecretAad(scope, id));
+  decipher.setAuthTag(Buffer.from(encodedTag, "base64url"));
+  return Buffer.concat([
+    decipher.update(Buffer.from(encodedCiphertext, "base64url")),
+    decipher.final(),
+  ]).toString("utf8");
+}
+
+function vaultSecretAad(scope: AccessScope, id: string) {
+  return Buffer.from(`${scope.workspaceId}\u0000vault\u0000${id}`);
 }
