+++ src/app/(authenticated)/(manager)/vault/_components/contacts/index.tsx
@@ -1,0 +1,83 @@
+"use client";
+
+import { PlusIcon } from "lucide-react";
+import type { VaultItem } from "@/lib/vault";
+import { Button } from "@/components/ui/button";
+import {
+  DialogDescription,
+  DialogHeader,
+  DialogTitle,
+} from "@/components/ui/dialog";
+import { ContactForm } from "./form";
+import {
+  useVaultSection,
+  VaultItemBrowser,
+  VaultSection,
+  VaultSectionBackButton,
+  VaultSectionContent,
+  VaultSectionTrigger,
+} from "../section";
+import { useVaultSetup } from "../setup";
+
+export function VaultContacts({
+  items,
+}: {
+  readonly items: readonly VaultItem[];
+}) {
+  const setup = useVaultSetup();
+  const initialAdd = setup?.kind === "contact";
+  const section = useVaultSection(initialAdd ? "add" : "list");
+
+  return (
+    <VaultSection
+      onOpenChange={section.onOpenChange}
+      open={section.open}
+      title="Contact info"
+    >
+      <VaultSectionTrigger items={items} title="Contact info" />
+      <VaultSectionContent view={section.view}>
+        {section.view === "list" ? (
+          <>
+            <DialogHeader className="pr-10 sm:pr-6">
+              <DialogTitle>Contact info</DialogTitle>
+              <DialogDescription>
+                {items.length > 0
+                  ? `Search and manage ${items.length.toLocaleString()} saved contact info.`
+                  : "Add your first saved contact."}
+              </DialogDescription>
+            </DialogHeader>
+            <VaultItemBrowser
+              items={items}
+              searchId="vault-search-contacts"
+              title="Contact info"
+            />
+            <div className="flex justify-end gap-2">
+              <Button onClick={() => section.setView("add")} type="button">
+                <PlusIcon />
+                Add contact
+              </Button>
+            </div>
+          </>
+        ) : (
+          <>
+            <VaultSectionBackButton
+              onClick={() => section.setView("list")}
+              title="Contact info"
+            />
+            <DialogHeader className="pr-10 sm:pr-6">
+              <DialogTitle>Add contact</DialogTitle>
+              <DialogDescription>
+                Sensitive values are encrypted before database storage and are
+                never returned after saving.
+              </DialogDescription>
+            </DialogHeader>
+            <ContactForm
+              initialLabel={initialAdd ? setup.label : undefined}
+              onSaved={() => section.setView("list")}
+            />
+          </>
+        )}
+      </VaultSectionContent>
+    </VaultSection>
+  );
+}
