+++ src/app/(authenticated)/(manager)/vault/_components/logins/import.tsx
@@ -7,6 +7,7 @@ import {
   UploadIcon,
 } from "lucide-react";
 import { useState } from "react";
+import { useRouter } from "next/navigation";
 import { Alert, AlertDescription, AlertTitle } from "@/components/ui/alert";
 import { Button } from "@/components/ui/button";
 import {
@@ -17,21 +18,19 @@ import {
 } from "@/components/ui/dialog";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
-import type { ManagerMutation } from "@/modules/manager";
-import { parseChromePasswordsCsv } from "@/modules/manager/chrome-passwords";
+import {
+  loginIdentifierSchema,
+  serializeLoginVaultPayload,
+  type VaultImportItems,
+} from "@/lib/vault";
+import { api } from "@/trpc/client";
 
 const MAX_FILE_SIZE = 10 * 1024 * 1024;
 const GOOGLE_PASSWORD_MANAGER_URL = "https://passwords.google.com/options";
 
-export function ChromePasswordImportPanel({
-  busy,
-  onDone,
-  onImport,
-}: {
-  readonly busy: boolean;
-  readonly onDone: () => void;
-  readonly onImport: (mutation: ManagerMutation) => Promise<boolean>;
-}) {
+export function ChromeImportPanel({ onDone }: { readonly onDone: () => void }) {
+  const router = useRouter();
+  const importPasswords = api.vault.import.useMutation();
   const [selection, setSelection] =
     useState<ReturnType<typeof parseChromePasswordsCsv>>();
   const [fileName, setFileName] = useState("");
@@ -40,6 +39,7 @@ export function ChromePasswordImportPanel({
   const [inputKey, setInputKey] = useState(0);
 
   const chooseFile = async (file?: File) => {
+    importPasswords.reset();
     setError(undefined);
     setImportedCount(undefined);
     setSelection(undefined);
@@ -61,34 +61,34 @@ export function ChromePasswordImportPanel({
     }
   };
 
-  const importPasswords = async () => {
+  const importSelectedPasswords = () => {
     if (!selection) return;
     setError(undefined);
     const count = selection.items.length;
-    const saved = await onImport({
-      action: "vault.import",
-      items: selection.items,
+    importPasswords.mutate(selection.items, {
+      onSuccess: () => {
+        router.refresh();
+        setSelection(undefined);
+        setImportedCount(count);
+        setFileName("");
+        setInputKey((key) => key + 1);
+      },
     });
-    if (!saved) {
-      setError(
-        "The import did not finish. Check the vault error and try again."
-      );
-      return;
-    }
-
-    setSelection(undefined);
-    setImportedCount(count);
-    setFileName("");
-    setInputKey((key) => key + 1);
   };
 
   const reset = () => {
+    importPasswords.reset();
     setSelection(undefined);
     setFileName("");
     setError(undefined);
     setImportedCount(undefined);
     setInputKey((key) => key + 1);
   };
+  const importError =
+    error ??
+    (importPasswords.error
+      ? "The import did not finish. Check the vault error and try again."
+      : undefined);
 
   return (
     <>
@@ -131,7 +131,7 @@ export function ChromePasswordImportPanel({
             </Label>
             <Input
               accept=".csv,text/csv"
-              disabled={busy}
+              disabled={importPasswords.isPending}
               id="chrome-passwords-csv"
               key={inputKey}
               onChange={(event) =>
@@ -150,11 +150,11 @@ export function ChromePasswordImportPanel({
             ) : null}
           </div>
 
-          {error ? (
+          {importError ? (
             <Alert variant="destructive">
               <FileKeyIcon />
               <AlertTitle>Couldn&apos;t import this file</AlertTitle>
-              <AlertDescription>{error}</AlertDescription>
+              <AlertDescription>{importError}</AlertDescription>
             </Alert>
           ) : null}
 
@@ -185,12 +185,12 @@ export function ChromePasswordImportPanel({
       <DialogFooter>
         {importedCount === undefined ? (
           <Button
-            disabled={busy || !selection}
-            onClick={() => void importPasswords()}
+            disabled={importPasswords.isPending || !selection}
+            onClick={importSelectedPasswords}
             type="button"
           >
             <UploadIcon />
-            {busy
+            {importPasswords.isPending
               ? "Importing…"
               : selection
                 ? `Import ${selection.items.length.toLocaleString()} ${selection.items.length === 1 ? "login" : "logins"}`
@@ -210,4 +210,153 @@ export function ChromePasswordImportPanel({
       </DialogFooter>
     </>
   );
+}
+
+function parseChromePasswordsCsv(csv: string) {
+  const rows = parseCsv(csv);
+  const headers = rows.shift()?.map((header) =>
+    header
+      .replace(/^\uFEFF/, "")
+      .trim()
+      .toLowerCase()
+  );
+  if (!headers) throw new Error("Choose a Chrome passwords CSV file.");
+
+  const indexes = {
+    name: headers.indexOf("name"),
+    password: headers.indexOf("password"),
+    url: headers.indexOf("url"),
+    username: headers.indexOf("username"),
+  };
+  if (indexes.url < 0 || indexes.username < 0 || indexes.password < 0) {
+    throw new Error(
+      "This CSV needs url, username, and password columns. Export it from Google Password Manager and try again."
+    );
+  }
+
+  const items: VaultImportItems = [];
+  let skipped = 0;
+
+  for (const row of rows) {
+    if (row.every((value) => value.length === 0)) continue;
+
+    const account = row[indexes.username]?.trim() ?? "";
+    const password = row[indexes.password] ?? "";
+    const url = row[indexes.url]?.trim() ?? "";
+    const origin = originFromUrl(url);
+    const name = indexes.name >= 0 ? row[indexes.name]?.trim() : undefined;
+    const label = name?.length ? name : labelFromUrl(url);
+
+    if (
+      !label ||
+      !origin ||
+      account.length === 0 ||
+      password.length === 0 ||
+      account.length > 300 ||
+      label.length > 120 ||
+      password.length > 20_000
+    ) {
+      skipped += 1;
+      continue;
+    }
+
+    items.push({
+      account: "",
+      kind: "login",
+      label,
+      secret: serializeLoginVaultPayload({
+        authentication: { password, type: "password" },
+        identifier: {
+          type: loginIdentifierSchema.safeParse({
+            type: "email",
+            value: account,
+          }).success
+            ? "email"
+            : "username",
+          value: account,
+        },
+        kind: "login",
+        origin,
+        version: 2,
+      }),
+    });
+  }
+
+  if (items.length === 0) {
+    throw new Error("No valid saved passwords were found in this CSV.");
+  }
+  if (items.length > 3_000) {
+    throw new Error(
+      `This file contains ${items.length.toLocaleString()} passwords. Import up to 3,000 at a time.`
+    );
+  }
+
+  return { items, skipped };
+}
+
+function labelFromUrl(value: string) {
+  if (!value) return "";
+  try {
+    const url = new URL(value);
+    return url.hostname.replace(/^www\./, "") || value;
+  } catch {
+    return value.slice(0, 120);
+  }
+}
+
+function originFromUrl(value: string) {
+  try {
+    const url = new URL(value);
+    return url.protocol === "http:" || url.protocol === "https:"
+      ? url.origin
+      : undefined;
+  } catch {
+    return undefined;
+  }
+}
+
+function parseCsv(csv: string) {
+  const rows: string[][] = [];
+  let field = "";
+  let quoted = false;
+  let row: string[] = [];
+
+  for (let index = 0; index < csv.length; index += 1) {
+    const character = csv.charAt(index);
+    if (quoted) {
+      if (character === '"') {
+        if (csv[index + 1] === '"') {
+          field += '"';
+          index += 1;
+        } else {
+          quoted = false;
+        }
+      } else {
+        field += character;
+      }
+      continue;
+    }
+
+    if (character === '"' && field.length === 0) {
+      quoted = true;
+    } else if (character === ",") {
+      row.push(field);
+      field = "";
+    } else if (character === "\n" || character === "\r") {
+      if (character === "\r" && csv[index + 1] === "\n") index += 1;
+      row.push(field);
+      rows.push(row);
+      field = "";
+      row = [];
+    } else {
+      field += character;
+    }
+  }
+
+  if (quoted) throw new Error("This CSV has an unfinished quoted value.");
+  if (field.length > 0 || row.length > 0) {
+    row.push(field);
+    rows.push(row);
+  }
+  return rows;
 }
