+++ src/app/(authenticated)/(manager)/vault/_components/setup.ts
@@ -1,0 +1,12 @@
+"use client";
+
+import { useSearchParams } from "next/navigation";
+import { parseVaultSetupSearchParams } from "@/lib/vault";
+
+export function useVaultSetup() {
+  const searchParams = useSearchParams();
+  const requestedSetup = parseVaultSetupSearchParams(
+    Object.fromEntries(searchParams.entries())
+  );
+  return requestedSetup.success ? requestedSetup.data : undefined;
+}
