--- src/modules/manager/vault-autofill.ts
@@ -1,6 +1,0 @@
-import { z } from "zod";
-
-export const fillFromVaultRequestSchema = z.object({
-  browserSessionId: z.string().trim().min(1).max(500),
-  candidateId: z.string().trim().min(1).max(500),
-});
