+++ tests/google-workspace.test.ts
@@ -1,10 +1,4 @@
-import {
-  getTokenResponse,
-  NoValidTokenError,
-  type ConnectTokenResponse,
-} from "@vercel/connect";
-import type * as VercelConnect from "@vercel/connect";
-import { afterEach, describe, expect, it, vi } from "vitest";
+import { describe, expect, it } from "vitest";
 import { parseCalendarAvailability } from "@/agent/lib/google-workspace/calendar";
 import { googleWorkspaceAuthOptions } from "@/agent/lib/google-workspace/client";
 import { gmailUpdateLabels } from "@/agent/lib/google-workspace/gmail";
@@ -14,34 +8,16 @@ import {
   googleWorkspaceSubject,
   googleWorkspaceTokenParams,
 } from "@/lib/google-workspace";
-import { readManagerSnapshot } from "@/modules/manager/server/store";
 
-vi.mock("@vercel/connect", async (importOriginal) => ({
-  ...(await importOriginal<typeof VercelConnect>()),
-  getTokenResponse: vi.fn<typeof getTokenResponse>(),
-}));
-vi.mock("@/db/services/settings", () => ({
-  getGatewayModel: vi.fn<() => Promise<string>>().mockResolvedValue("test"),
-  selectGatewayModel: vi.fn<() => Promise<void>>(),
-}));
-vi.mock("@/modules/manager/server/vault", () => ({
-  readManagerVaultItems: vi.fn<() => Promise<never[]>>().mockResolvedValue([]),
-}));
+const userId = "better-auth:user-123";
 
-afterEach(() => vi.clearAllMocks());
-
-const scope = {
-  userId: "better-auth:user-123",
-  workspaceId: "personal:workspace-123",
-};
-
-describe("Google Workspace connection", () => {
+describe("Google Workspace", () => {
   it("uses one explicit least-privilege scope set", () => {
     expect(googleWorkspaceScopes).not.toContain("*");
     expect(googleWorkspaceScopes).not.toContain("https://mail.google.com/");
-    expect(googleWorkspaceTokenParams(scope.userId)).toEqual({
+    expect(googleWorkspaceTokenParams(userId)).toEqual({
       scopes: [...googleWorkspaceScopes],
-      subject: googleWorkspaceSubject(scope.userId),
+      subject: googleWorkspaceSubject(userId),
     });
     expect(googleWorkspaceAuthOptions.tokenParams).toEqual({
       scopes: [...googleWorkspaceScopes],
@@ -49,46 +25,15 @@ describe("Google Workspace connection", () => {
     expect(googleWorkspaceAuthOptions.validate).toBe(true);
   });
 
-  it("uses the same user subject for manager and Eve flows", () => {
-    expect(googleWorkspaceSubject(scope.userId)).toEqual({
-      id: scope.userId,
+  it("uses a user-scoped connector subject", () => {
+    expect(googleWorkspaceSubject(userId)).toEqual({
+      id: userId,
       issuer: "openinstinct",
       type: "user",
     });
   });
 
-  it("reports connected accounts without exposing tokens", async () => {
-    const response: ConnectTokenResponse = {
-      claims: { email: "person@example.com" },
-      connector: { id: "connector-id", type: "oauth", uid: "google/test" },
-      expiresAt: Date.now() + 60_000,
-      token: "must-not-leak",
-    };
-    vi.mocked(getTokenResponse).mockResolvedValue(response);
-
-    await expect(readManagerSnapshot(scope)).resolves.toMatchObject({
-      googleWorkspace: {
-        accountLabel: "person@example.com",
-        state: "connected",
-      },
-    });
-    expect(getTokenResponse).toHaveBeenCalledWith(
-      expect.any(String),
-      googleWorkspaceTokenParams(scope.userId),
-      { forceRefresh: true }
-    );
-  });
-
-  it("reports a missing user grant as disconnected", async () => {
-    vi.mocked(getTokenResponse).mockRejectedValue(
-      new NoValidTokenError("No Google grant for this user.")
-    );
-    await expect(readManagerSnapshot(scope)).resolves.toMatchObject({
-      googleWorkspace: { accountLabel: null, state: "disconnected" },
-    });
-  });
-
-  it("maps reversible Gmail actions to system labels", () => {
+  it("maps reversible Gmail actions and protects consequential writes", () => {
     expect(gmailUpdateLabels("archive")).toEqual({
       addLabelIds: [],
       removeLabelIds: ["INBOX"],
@@ -97,17 +42,11 @@ describe("Google Workspace connection", () => {
       addLabelIds: ["UNREAD"],
       removeLabelIds: [],
     });
-  });
-
-  it("requires approval for consequential writes only", () => {
     expect(googleWorkspaceWriteApproval("update_email")).toBe("not-applicable");
     expect(googleWorkspaceWriteApproval("send_email")).toBe("user-approval");
-    expect(googleWorkspaceWriteApproval("create_calendar_event")).toBe(
-      "user-approval"
-    );
   });
 
-  it("does not interpret Google FreeBusy errors as availability", () => {
+  it("does not treat calendar API errors as availability", () => {
     expect(() =>
       parseCalendarAvailability({
         calendars: {
@@ -117,31 +56,5 @@ describe("Google Workspace connection", () => {
         },
       })
     ).toThrow(/missing@example\.com: notFound/u);
-
-    expect(
-      parseCalendarAvailability({
-        calendars: {
-          primary: {
-            busy: [
-              {
-                end: "2026-08-27T15:00:00-04:00",
-                start: "2026-08-27T14:00:00-04:00",
-              },
-            ],
-          },
-        },
-      })
-    ).toEqual({
-      calendars: {
-        primary: {
-          busy: [
-            {
-              end: "2026-08-27T15:00:00-04:00",
-              start: "2026-08-27T14:00:00-04:00",
-            },
-          ],
-        },
-      },
-    });
   });
 });
