+++ tests/vault-autofill.test.ts
@@ -1,28 +1,41 @@
-import { describe, expect, it } from "vitest";
+import { describe, expect, it, vi } from "vitest";
 import type { AccessScope } from "@/lib/access-scope";
-import type { VaultItemKind } from "@/modules/manager";
-import { serializePaymentCard } from "@/modules/manager/payment-card";
-import {
-  classifyNativeLoginControl,
-  selectNativeLoginFills,
-  type NativeLoginControlDescriptor,
-} from "@/modules/manager/server/kernel-login-autofill";
-import {
-  buildNativeAutofillPayload,
-  nativeAutofillSecretMarkingExpression,
-  nativeAutofillTokens,
-} from "@/modules/manager/server/kernel-native-autofill";
-import {
-  listAutofillSuggestions,
-  materializeAutofillClaims,
-  type AutofillVaultAdapter,
-} from "@/modules/manager/server/vault-autofill";
-import { createVaultAutofillProvider } from "@/modules/manager/server/vault-autofill-provider";
 import {
   serializeAddressVaultPayload,
   serializeContactVaultPayload,
   serializeLoginVaultPayload,
-} from "@/modules/manager/vault-payload";
+  serializePaymentCard,
+  type VaultItemKind,
+} from "@/lib/vault";
+import {
+  classifyNativeLoginControl,
+  selectNativeLoginFills,
+  type NativeLoginControlDescriptor,
+} from "@/agent/subagents/worker/lib/autofill/login";
+import {
+  buildNativeAutofillPayload,
+  nativeAutofillSecretMarkingExpression,
+  nativeAutofillTokens,
+} from "@/agent/subagents/worker/lib/autofill/native";
+import {
+  listAutofillSuggestions,
+  materializeAutofillClaims,
+  type AutofillVaultAdapter,
+} from "@/agent/subagents/worker/lib/autofill/service";
+import { vaultAutofillProvider } from "@/agent/subagents/worker/lib/autofill/provider";
+
+const vaultStore = vi.hoisted(() => ({
+  items: [] as { id: string }[],
+  secret: "",
+}));
+
+vi.mock("@/db/services/vault", () => ({
+  hasVaultSecret: async () => true,
+  listVaultItems: async () => vaultStore.items,
+  readVaultItem: async (_scope: unknown, id: string) =>
+    vaultStore.items.find((item) => item.id === id),
+  readVaultSecret: async () => vaultStore.secret,
+}));
 
 const scope: AccessScope = {
   userId: "user-1",
@@ -63,29 +76,19 @@ describe("vault browser autofill", () => {
       label: "Travel card",
       updatedAt: "2026-08-27T00:00:00.000Z",
     };
-    const provider = createVaultAutofillProvider({
-      async hasSecret() {
-        return true;
-      },
-      async listVaultItems() {
-        return [card];
-      },
-      async readSecret() {
-        return serializePaymentCard({
-          billingPostalCode: "10001",
-          cardholderName: "Grace Hopper",
-          expirationMonth: 9,
-          expirationYear: 2031,
-          kind: "payment-card",
-          number: "4111111111111111",
-          securityCode: "321",
-          version: 1,
-        });
-      },
-      async readVaultItem() {
-        return card;
-      },
-    });
+    const provider = providerFor(
+      card,
+      serializePaymentCard({
+        billingPostalCode: "10001",
+        cardholderName: "Grace Hopper",
+        expirationMonth: 9,
+        expirationYear: 2031,
+        kind: "payment-card",
+        number: "4111111111111111",
+        securityCode: "321",
+        version: 1,
+      })
+    );
 
     await expect(
       provider.listSuggestions(
@@ -605,20 +608,9 @@ function vaultItem(kind: VaultItemKind, label: string, account: string) {
 }
 
 function providerFor(item: ReturnType<typeof vaultItem>, secret: string) {
-  return createVaultAutofillProvider({
-    async hasSecret() {
-      return true;
-    },
-    async listVaultItems() {
-      return [item];
-    },
-    async readSecret() {
-      return secret;
-    },
-    async readVaultItem() {
-      return item;
-    },
-  });
+  vaultStore.items = [item];
+  vaultStore.secret = secret;
+  return vaultAutofillProvider;
 }
 
 function claimValues(
